backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Backend service for the local Linux RAG CLIs."
requires-python = ">=3.12"
dependencies = [
    "structlog",
    "weaviate-client",
    "arize-phoenix",
    "arize-phoenix-otel>=0.13.1",
    "pyyaml>=6.0",
    "openinference-instrumentation>=0.1.41",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
]


[dependency-groups]
dev = [
    "build>=1.2.2",
    "ruff>=0.14.4",
    "mypy>=1.18.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=7.0.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.23",
    "types-pyyaml>=6.0.12.20250915",
    "mkdocstrings-python>=1.19.0",
    "mkdocs-autoapi>=0.4.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.distutils.egg_info]
egg_base = "build"

[tool.mypy]
mypy_path = ["src"]
explicit_package_bases = true
